requirements.txt
fastapi
uvicorn
python-pptx
convertapi
supabase
python-dotenv
